push complexity of "maybe the files are not there" to external script
Each line of the CodeBuild buildspec is apparently pushed into a script
then run with an exec in the container. shopts get reset with every line,
but there also seems some weirdness around quoting that I have not gotten
my head around, but is avoided by just running a script.